fix: free in search_by_key_in_table

+++ 3b/src/client/main.c
@@ -13,7 +13,7 @@ int main(int argc, char **argv){
         return 1;
     }
 
-    size_t atoi_argv_1 = atoi(argv[1]);
+    int atoi_argv_1 = atoi(argv[1]);
     if (atoi_argv_1 <= 0) {
         printf("Error: table size mismatch\n");
         return 1;
